@@ -10,32 +10,35 @@
   "license": "GPL-3",
   "dependencies": {
     "axios": "^0.18.0",
+    "babel-core": "^6.26.3",
     "babel-loader": "^7.1.4",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "chokidar": "^2.0.3",
     "clean-webpack-plugin": "^0.1.19",
     "docker": "^1.0.0",
     "express": "^4.16.3",
     "find-process": "^1.1.1",
     "html-webpack-plugin": "^3.2.0",
+    "jshint-loader": "^0.8.4",
     "loadash": "^1.0.0",
     "loader": "^2.1.1",
+    "lodash": "^4.17.10",
     "material-ui": "^0.20.1",
+    "material-ui-table-edit": "^3.0.3",
     "modules": "^0.4.0",
     "react": "^16.4.0",
     "react-dom": "^16.4.0",
     "react-router": "^4.3.1",
     "redux": "^4.0.0",
     "redux-thunk": "^2.3.0",
-    "transpile": "babel app/app.jsx --out-file app/app-transpiled.js",
+    "ts-loader": "^4.4.1",
     "uuid": "^3.2.1",
-    "webpack": "^4.11.1",
+    "webpack": "^4.12.0",
+    "webpack-dev-server": "^3.1.4",
     "webpack-hot-middleware": "^2.22.2"
   },
-  "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "ts-loader": "^4.4.0"
-  }
+  "devDependencies": {}
 }
